add backend swithc for functaional APIs
