+++ fusionlab/functional/__init__.py
@@ -1,8 +1,15 @@
-from .dice import dice_score, tf_dice_score
-from .iou import (
-    iou_score,
-    tf_iou_score,
-    jaccard_score,
-    tf_jaccard_score,
-)
-
+from fusionlab.configs import BACKEND
+if BACKEND == 'torch':
+    from .dice import dice_score
+    from .iou import (
+        iou_score,
+        jaccard_score,
+    )
+elif BACKEND == 'tf':
+    from .dice import tf_dice_score
+    from .iou import (
+        tf_iou_score,
+        tf_jaccard_score,
+    )
+else:
+    print('backend not supported!!!')
